[refactor] 역할별 기능 수정 - page 에서 데이터 가져오는 기능 container 로 수정
