+++ src/containers/SeriesContainer.test.jsx
@@ -1,19 +1,37 @@
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+
+import { getDefaultMiddleware } from '@reduxjs/toolkit';
+
+import configureStore from 'redux-mock-store';
 
 import { render } from '@utils/test-utils';
+
+import {
+  loadSeries,
+  setSeries,
+} from '@modules/slice';
 
 import SERIES from '@/fixture/series';
 
 import SeriesContainer from './SeriesContainer';
 
+const mockStore = configureStore(getDefaultMiddleware());
+
 jest.mock('react-redux');
 jest.mock('@libs/api');
 
 describe('PostsContainer', () => {
+  let store;
+
+  const dispatch = jest.fn();
+
   beforeEach(() => {
+    dispatch.mockClear();
+
     useSelector.mockImplementation((selector) => selector({
       series: given.series,
     }));
+    useDispatch.mockImplementation(() => dispatch);
   });
 
   function renderSeriesContainer() {
@@ -21,6 +39,28 @@ describe('PostsContainer', () => {
       <SeriesContainer />,
     );
   }
+
+  context('when load', () => {
+    given('series', () => []);
+
+    beforeEach(() => {
+      store = mockStore({});
+    });
+
+    it('dispatchs loadSeries', async () => {
+      renderSeriesContainer();
+
+      expect(dispatch).toBeCalled();
+
+      await store.dispatch(loadSeries());
+
+      const actions = store.getActions();
+
+      expect(actions).toEqual([
+        setSeries([]),
+      ]);
+    });
+  });
 
   context('with series', () => {
     given('series', () => SERIES);
