+++ src/pages/SeriesPage.test.jsx
@@ -1,35 +1,17 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import { getDefaultMiddleware } from '@reduxjs/toolkit';
-
-import configureStore from 'redux-mock-store';
-
 import { render } from '@utils/test-utils';
 
-import {
-  loadSeries,
-  setSeries,
-} from '@modules/slice';
+import SERIES from '@/fixture/series';
 
 import SeriesPage from './SeriesPage';
 
-const mockStore = configureStore(getDefaultMiddleware());
-
-jest.mock('react-redux');
-jest.mock('@libs/api');
-
 describe('SeriesPage', () => {
-  let store;
-
-  const dispatch = jest.fn();
-
   beforeEach(() => {
-    dispatch.mockClear();
-
+    useDispatch.mockImplementation(() => jest.fn());
     useSelector.mockImplementation((selector) => selector({
-      series: [],
+      series: SERIES,
     }));
-    useDispatch.mockImplementation(() => dispatch);
   });
 
   function renderSeriesPage() {
@@ -38,26 +20,12 @@ describe('SeriesPage', () => {
     );
   }
 
-  context('when load', () => {
-    beforeEach(() => {
-      store = mockStore({});
-    });
+  it('renders series', () => {
+    const { container } = renderSeriesPage();
 
-    it('dispatchs loadSeries', async () => {
-      renderSeriesPage();
-
-      expect(dispatch).toBeCalled();
-
-      await store.dispatch(loadSeries());
-
-      const actions = store.getActions();
-
-      expect(actions).toEqual([
-        setSeries([]),
-      ]);
+    SERIES.forEach(({ title, subTitle }) => {
+      expect(container).toHaveTextContent(title);
+      expect(container).toHaveTextContent(subTitle);
     });
   });
-
-  // TODO: series 가 있는 경우, 없는 경우 테스트 ?
-  // 컨테이너에서 검사하는데 굳이 해야할까?
 });
